refactor: move servers composables to `@modrinth/ui`
