+++ apps/frontend/src/pages/servers/manage/[id]/options/info.vue
@@ -117,7 +117,7 @@
 <script setup lang="ts">
 import { ButtonStyled } from "@modrinth/ui";
 import { CopyIcon, ExternalIcon, EyeIcon, EyeOffIcon } from "@modrinth/assets";
-import { ModrinthServer } from "~/composables/servers/modrinth-servers.ts";
+import { ModrinthServer } from "@modrinth/ui";
 
 const route = useNativeRoute();
 const serverId = route.params.id as string;
