+++ apps/frontend/src/pages/servers/manage/[id]/options/preferences.vue
@@ -43,7 +43,7 @@
 
 <script setup lang="ts">
 import { useStorage } from "@vueuse/core";
-import { ModrinthServer } from "~/composables/servers/modrinth-servers.ts";
+import { ModrinthServer } from "@modrinth/ui";
 
 const route = useNativeRoute();
 const serverId = route.params.id as string;
