+++ apps/frontend/src/pages/settings/billing/index.vue
@@ -591,7 +591,7 @@ import {
 } from "@modrinth/assets";
 import { calculateSavings, formatPrice, getCurrency } from "@modrinth/utils";
 import { ref, computed } from "vue";
-import { useServersFetch } from "~/composables/servers/servers-fetch.ts";
+import { useServersFetch } from "@modrinth/ui";
 import { products } from "~/generated/state.json";
 
 definePageMeta({
