+++ packages/ui/src/composables/servers/modules/content.ts
@@ -1,6 +1,6 @@
 import type { Mod, ContentType } from "@modrinth/utils";
-import { useServersFetch } from "../servers-fetch.ts";
-import { ServerModule } from "./base.ts";
+import { useServersFetch } from "../servers-fetch.js";
+import { ServerModule } from "./base.js";
 
 export class ContentModule extends ServerModule {
   data: Mod[] = [];
