+++ packages/ui/src/composables/servers/modules/fs.ts
@@ -6,8 +6,8 @@ import type {
   FSQueuedOp,
 } from "@modrinth/utils";
 import { ModrinthServerError } from "@modrinth/utils";
-import { useServersFetch } from "../servers-fetch.ts";
-import { ServerModule } from "./base.ts";
+import { useServersFetch } from "../servers-fetch.js";
+import { ServerModule } from "./base.js";
 
 export class FSModule extends ServerModule {
   auth!: JWTAuth;
