+++ packages/ui/src/composables/servers/modules/network.ts
@@ -1,6 +1,6 @@
 import type { Allocation } from "@modrinth/utils";
-import { useServersFetch } from "../servers-fetch.ts";
-import { ServerModule } from "./base.ts";
+import { useServersFetch } from "../servers-fetch.js";
+import { ServerModule } from "./base.js";
 
 export class NetworkModule extends ServerModule {
   allocations: Allocation[] = [];
