+++ packages/ui/src/composables/servers/modules/startup.ts
@@ -1,6 +1,6 @@
 import type { Startup, JDKVersion, JDKBuild } from "@modrinth/utils";
-import { useServersFetch } from "../servers-fetch.ts";
-import { ServerModule } from "./base.ts";
+import { useServersFetch } from "../servers-fetch.js";
+import { ServerModule } from "./base.js";
 
 export class StartupModule extends ServerModule implements Startup {
   invocation!: string;
