+++ packages/ui/src/composables/servers/modules/ws.ts
@@ -1,6 +1,6 @@
 import type { JWTAuth } from "@modrinth/utils";
-import { useServersFetch } from "../servers-fetch.ts";
-import { ServerModule } from "./base.ts";
+import { useServersFetch } from "../servers-fetch.js";
+import { ServerModule } from "./base.js";
 
 export class WSModule extends ServerModule implements JWTAuth {
   url!: string;
